+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Improved text snapping to trim some symbols in the end of selection (,:")
 - Fixed performance leak when clicking inside text field
 - Improved dark page detection
+- Added option "Add 'Paste' button only when field is empty"
+- Set higher z-index for better visibility on websites
 - Fixed dragging tooltip by arrow
 
 #3.3.3